+++ v2/components/icons/ChevronUp/ChevronUp.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 
 interface ChevronUpProps extends IconProps {
