+++ v2/components/icons/EthereumIcon/EthereumIcon.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Icon } from '@chakra-ui/react';
 
 interface EthereumIconProps {
