+++ v2/components/icons/InfoIcon/InfoIcon.tsx
@@ -1,11 +1,7 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 
-interface InfoIconProps extends IconProps {
-  width?: string | number;
-  height?: string | number;
-}
-
-export const InfoIcon = ({ width = '20px', height = '20px', color = 'white' }: InfoIconProps) => {
+export const InfoIcon = ({ width = '20px', height = '20px', color = 'white' }: IconProps) => {
   return (
     <Icon width={width} height={height} viewBox="0 0 24 24" fill="none" color={color}>
       <path
