+++ v2/components/icons/LoansIcon/LoansIcon.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 
 export const LoansIcon = ({
