+++ v2/components/icons/MaintainIcon/MaintainIcon.tsx
@@ -1,36 +1,24 @@
 import React from 'react';
 import { IconProps, Icon } from '@chakra-ui/react';
 
-interface MaintainIconProps extends IconProps {
-  width?: number;
-  height?: number;
-  color?: string;
-}
 export const MaintainIcon = ({
-  width = 28,
-  height = 36,
+  width = '28px',
+  height = '36px',
   color = '#FF9A54',
-
   ...props
-}: MaintainIconProps) => (
-  <Icon
-    width={`${width}px`}
-    height={`${height}px`}
-    fill="none"
-    viewBox={`0 0 ${width} ${height}`}
-    {...props}
-  >
+}: IconProps) => (
+  <Icon width={width} height={height} fill="none" viewBox="0 0 28 36" {...props} color={color}>
     <path
       fillRule="evenodd"
       clipRule="evenodd"
       d="m14.903 15.297-.904-.856-.902.857a16.31 16.31 0 0 0-5.044 10.734l-.003.044v.044c0 1.816.612 3.357 1.735 4.442 1.115 1.076 2.62 1.595 4.215 1.595 1.58 0 3.087-.492 4.21-1.548 1.138-1.069 1.758-2.598 1.758-4.42v-.04l-.003-.042a16.31 16.31 0 0 0-5.062-10.81Zm-4.228 10.865a13.685 13.685 0 0 1 3.326-8.018 13.686 13.686 0 0 1 3.342 8.086c-.01 1.183-.401 1.97-.93 2.466-.547.515-1.367.836-2.413.836-1.03 0-1.843-.33-2.391-.859-.533-.514-.925-1.323-.934-2.51Z"
-      fill={color}
+      fill="currentColor"
     />
     <path
       fillRule="evenodd"
       clipRule="evenodd"
       d="m14 .759-5.454 8.18L4.86 5.255l-.797 1.685c-.57 1.207-3.28 7.966-3.37 14.514-.044 3.281.566 6.704 2.639 9.327C5.45 33.459 8.9 35.062 14 35.062c5.101 0 8.552-1.61 10.67-4.296 2.071-2.629 2.681-6.06 2.636-9.345-.088-6.557-2.796-13.317-3.378-14.5l-.81-1.646-3.664 3.665L14 .759Zm0 4.732 5.046 7.57 3.237-3.238c.924 2.51 2.338 7.165 2.399 11.633.04 3.006-.534 5.73-2.074 7.685-1.495 1.896-4.085 3.296-8.608 3.296-4.525 0-7.115-1.395-8.609-3.285-1.539-1.947-2.113-4.663-2.073-7.663.061-4.465 1.478-9.126 2.406-11.659l3.23 3.23L14 5.491Z"
-      fill={color}
+      fill="currentColor"
     />
   </Icon>
 );
