+++ v2/components/icons/NotificationsIcon/NotificationsIcon.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 
 export const NotificationsIcon = ({
