+++ v2/components/icons/SLINKIcon/SLINKIcon.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 
 interface SLINKIconProps extends IconProps {
