+++ v2/components/icons/SNXIcon/SNXIcon.tsx
@@ -1,21 +1,20 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 
 interface SnxIconProps extends IconProps {
-  width?: number;
-  height?: number;
   color?: string;
   backgroundColor?: string;
 }
 
 export const SNXIcon = ({
-  width = 25,
-  height = 24,
+  width = '25px',
+  height = '24px',
   color = '#2ED9FF',
   backgroundColor = '#012546',
   ...props
 }: SnxIconProps) => {
   return (
-    <Icon width={`${width}px`} height={`${height}px`} viewBox="0 0 25 24" fill="none" {...props}>
+    <Icon width={width} height={height} viewBox="0 0 25 24" fill="none" {...props}>
       <path
         d="M24.5 12C24.5 18.6274 19.1274 24 12.5 24C5.87258 24 0.5 18.6274 0.5 12C0.5 5.37258 5.87258 0 12.5 0C19.1274 0 24.5 5.37258 24.5 12Z"
         fill={backgroundColor}
