+++ v2/components/icons/SettingsIcon/SettingsIcon.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 
 export const SettingsIcon = ({
