+++ v2/components/icons/StakingIcon/StakingIcon.tsx
@@ -1,10 +1,7 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
 
-interface StakingIconProps extends IconProps {
-  color?: string;
-}
-
-export const StakingIcon = ({ color = 'cyan.400' }: StakingIconProps) => {
+export const StakingIcon = ({ color = 'cyan.400' }: IconProps) => {
   return (
     <Icon width="29" height="21" viewBox="0 0 29 21" fill="none" color={color}>
       <path
