+++ v2/components/icons/TriangleUpIcon/TriangleUpIcon.tsx
@@ -1,16 +1,12 @@
+import React from 'react';
 import { Icon, IconProps } from '@chakra-ui/react';
-
-interface TriangleUpIconProps extends IconProps {
-  width?: number | string;
-  height?: number | string;
-}
 
 export const TriangleUpIcon = ({
   width = '25px',
   height = '24px',
   color = 'white',
   ...props
-}: TriangleUpIconProps) => {
+}: IconProps) => {
   return (
     <Icon width={width} height={height} viewBox="0 0 25 24" fill="none" color={color} {...props}>
       <path d="M12.3564 8L18.3564 14H6.35645L12.3564 8Z" fill="currentColor" />
